refactor: PluginController and module
Removes unused imports and logic related to plugin creation, updating, activation, deactivation, installation, uninstallation, verification, and version management from `PluginController`.

Updates `PluginRegistryModule` imports and adds `RolePermissionModule`.

This change simplifies the controller and module by removing obsolete code.
