+++ packages/plugins/plugin-registry/src/lib/plugin-registry.module.ts
@@ -4,14 +4,20 @@ import { CqrsModule } from '@nestjs/cqrs';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { handlers } from './application/handlers';
 import { entities } from './domain/entities';
-import { services } from './domain/services';
-import { PluginController } from './infrastructure/controllers/plugin.controller';
 import { repositories } from './domain/repositories';
+import { services } from './domain/services';
+import { controllers } from './infrastructure/controllers';
 import { subscribers } from './infrastructure/subscribers';
+import { RolePermissionModule } from '@gauzy/core';
 
 @Module({
-	controllers: [PluginController],
-	imports: [TypeOrmModule.forFeature([...entities]), MikroOrmModule.forFeature([...entities]), CqrsModule],
-	providers: [...services, ...handlers, ...repositories, ...subscribers]
+	imports: [
+		TypeOrmModule.forFeature([...entities]),
+		MikroOrmModule.forFeature([...entities]),
+		RolePermissionModule,
+		CqrsModule
+	],
+	providers: [...services, ...handlers, ...repositories, ...subscribers],
+	controllers
 })
 export class PluginRegistryModule {}
